requirements.txt
absl-py==0.12.0
alembic==1.6.2
analytics-python==1.2.9
appdirs==1.4.4
argon2-cffi==20.1.0
astropy==4.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
audioread==2.1.9
avalanche-lib==0.1.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
bleach==3.3.0
boto3==1.21.4
botocore==1.24.4
cachetools==4.2.1
captum==0.4.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==8.0.4
cliff==3.7.0
cmaes==0.8.2
cmd2==1.5.0
colorama==0.4.4
colorlog==5.0.1
configs==3.0.3
confuse==1.4.0
contextvars==2.4
convertdate==2.3.2
cryptography==3.4.7
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
deep-translator==1.5.0
defusedxml==0.7.1
docker-pycreds==0.4.0
einops==0.4.1
emoji==1.7.0
entrypoints==0.3
et-xmlfile==1.1.0
ffmpy==0.3.0
filelock==3.0.12
Flask==1.1.2
Flask-BasicAuth==0.2.0
Flask-CacheBuster==1.0.0
Flask-Cors==3.0.10
flatbuffers==1.12
funcy==1.17
future==0.18.2
gast==0.3.3
gdown==4.2.1
gensim==4.1.2
gitdb==4.0.9
GitPython==3.1.18
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googletrans==4.0.0rc1
GPUtil==1.4.0
gradio==1.6.4
greenlet==1.1.0
grpcio==1.32.0
h11==0.9.0
h2==3.2.0
h5py==2.10.0
hijri-converter==2.2.4
holidays==0.13
hpack==3.0.0
hstspreload==2022.12.1
htmlmin==0.1.12
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.4.0
hyperframe==5.2.0
idna==2.10
ImageHash==4.2.0
imageio==2.9.0
immutables==0.19
importlib-metadata==4.0.0
iniconfig==1.1.1
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
korean-lunar-calendar==0.3.1
librosa==0.8.0
llvmlite==0.36.0
Mako==1.1.4
Markdown==3.3.4
markdown2==2.4.0
MarkupSafe==1.1.1
matplotlib==3.3.4
missingno==0.4.2
missingpy==0.2.0
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nltk==3.6.1
notebook==6.3.0
numba==0.53.1
numexpr==2.7.3
numpy==1.19.5
oauthlib==3.1.0
openpyxl==3.0.9
opt-einsum==3.3.0
optuna==2.7.0
orca==1.6
packaging==20.9
pandas==1.1.5
pandas-profiling==2.11.0
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.2
pathtools==0.1.2
patsy==0.5.3
pbr==5.6.0
pexpect==4.8.0
phik==0.11.2
pickle5==0.0.12
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.14.3
pluggy==1.0.0
pooch==1.3.0
prettytable==2.1.0
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.15.8
psrecord==1.2
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyboolnet @ git+https://github.com/hklarner/pyboolnet@0213035cd65f1d3f7ff0310a1c446c8b662e76a0
pycocotools==2.0.4
pycparser==2.20
Pygments==2.8.1
pyLDAvis==3.2.2
PyMeeus==0.5.12
PyNaCl==1.4.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.17.3
PySocks==1.7.1
pyspellchecker==0.7.2
pytest==7.0.1
python-dateutil==2.8.1
python-editor==1.0.4
pytorchcv==0.0.67
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
quadprog==0.1.11
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
retrying==1.3.3
rfc3986==1.5.0
rsa==4.7.2
s3transfer==0.5.1
sacremoses==0.0.45
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
sentencepiece==0.1.96
sentry-sdk==1.5.4
shortuuid==1.0.8
six==1.15.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
SoundFile==0.10.3.post1
soupsieve==2.3.1
SQLAlchemy==1.4.14
statsmodels==0.12.2
stevedore==3.3.0
ta==0.8.0
tab-transformer-pytorch==0.1.4
tables==3.6.1
tangled-up-in-unicode==0.0.7
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
tokenizers==0.12.1
tomli==1.2.3
toolz==0.11.1
torch==1.10.2
torchvision==0.11.2
tornado==6.1
tqdm==4.60.0
traitlets==4.3.3
transformers==4.18.0
typing-extensions==3.7.4.3
Unidecode==1.3.3
urllib3==1.26.4
visions==0.6.0
wandb==0.12.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
yaspin==2.1.0
zipp==3.4.1
